Split PlayerDataHandler into multiple classes, and then deprecated everything.
