+++ gradle.properties
@@ -1,15 +1,15 @@
-mod_version=1.0.
+mod_version=2.0.
 mod_group=net.darkhax.gamestages
 mod_name=GameStages
 mod_class=GameStages
 
 version_minecraft=1.12.2
-version_forge=14.23.2.2611
+version_forge=14.23.3.2673
 version_mcp=snapshot_20180204
 
 curse_versions=1.12.2
 curse_requirements=bookshelf
 curse_project=268655
 
-version_bookshelf=2.3.523
+version_bookshelf=2.3.544
 version_minetweaker=4.1.2.397
